refactor: move requirement class outside of solver
